NS1: Moves account permissions logic to permissions.go
