+++ builtin/providers/nsone/resource_user.go
@@ -7,88 +7,6 @@ import (
 	"gopkg.in/ns1/ns1-go.v2/rest/model/account"
 )
 
-func addPermsSchema(s map[string]*schema.Schema) map[string]*schema.Schema {
-	s["dns_view_zones"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["dns_manage_zones"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["dns_zones_allow_by_default"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["dns_zones_deny"] = &schema.Schema{
-		Type:     schema.TypeList,
-		Optional: true,
-		Elem:     &schema.Schema{Type: schema.TypeString},
-	}
-	s["dns_zones_allow"] = &schema.Schema{
-		Type:     schema.TypeList,
-		Optional: true,
-		Elem:     &schema.Schema{Type: schema.TypeString},
-	}
-	s["data_push_to_datafeeds"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["data_manage_datasources"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["data_manage_datafeeds"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_users"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_payment_methods"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_plan"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_teams"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_apikeys"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_manage_account_settings"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_view_activity_log"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["account_view_invoices"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["monitoring_manage_lists"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["monitoring_manage_jobs"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	s["monitoring_view_jobs"] = &schema.Schema{
-		Type:     schema.TypeBool,
-		Optional: true,
-	}
-	return s
-}
-
 func userResource() *schema.Resource {
 	s := map[string]*schema.Schema{
 		"id": &schema.Schema{
@@ -133,28 +51,6 @@ func userResource() *schema.Resource {
 		Update: UserUpdate,
 		Delete: UserDelete,
 	}
-}
-
-func permissionsToResourceData(d *schema.ResourceData, permissions account.PermissionsMap) {
-	d.Set("dns_view_zones", permissions.DNS.ViewZones)
-	d.Set("dns_manage_zones", permissions.DNS.ManageZones)
-	d.Set("dns_zones_allow_by_default", permissions.DNS.ZonesAllowByDefault)
-	d.Set("dns_zones_deny", permissions.DNS.ZonesDeny)
-	d.Set("dns_zones_allow", permissions.DNS.ZonesAllow)
-	d.Set("data_push_to_datafeeds", permissions.Data.PushToDatafeeds)
-	d.Set("data_manage_datasources", permissions.Data.ManageDatasources)
-	d.Set("data_manage_datafeeds", permissions.Data.ManageDatafeeds)
-	d.Set("account_manage_users", permissions.Account.ManageUsers)
-	d.Set("account_manage_payment_methods", permissions.Account.ManagePaymentMethods)
-	d.Set("account_manage_plan", permissions.Account.ManagePlan)
-	d.Set("account_manage_teams", permissions.Account.ManageTeams)
-	d.Set("account_manage_apikeys", permissions.Account.ManageApikeys)
-	d.Set("account_manage_account_settings", permissions.Account.ManageAccountSettings)
-	d.Set("account_view_activity_log", permissions.Account.ViewActivityLog)
-	d.Set("account_view_invoices", permissions.Account.ViewInvoices)
-	d.Set("monitoring_manage_lists", permissions.Monitoring.ManageLists)
-	d.Set("monitoring_manage_jobs", permissions.Monitoring.ManageJobs)
-	d.Set("monitoring_view_jobs", permissions.Monitoring.ViewJobs)
 }
 
 func userToResourceData(d *schema.ResourceData, u *account.User) error {
